+++ addons/yat/src/commands/builtin/Set.cs
@@ -1,3 +1,4 @@
+using System;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
@@ -13,13 +14,10 @@ public partial class Set : Extensible, ICommand
 {
 	public CommandResult Execute(CommandData data)
 	{
-		var variable = data.RawData[1];
+		var extensions = GetCommandExtensions("set");
 
-		if (Extensions.ContainsKey(variable))
-		{
-			var extension = Extensions[variable];
-			return extension.Execute(data with { RawData = data.RawData[1..] });
-		}
+		if (extensions.TryGetValue((string)data.Arguments["variable"], out Type extension))
+			return ExecuteExtension(extension, data with { RawData = data.RawData[1..] });
 
 		data.Terminal.Print("Variable not found.", PrintType.Error);
 		return CommandResult.Failure;